// tools/json-formatter/client/types.ts
export type JsonFormatterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  importFromFileLabel: string
  downloadJsonLabel: string
  rawJsonLabel: string
  rawJsonDescription: string
  formattedJsonLabel: string
  formattedJsonDescription: string
  jsonPlaceholder: string
  invalidJsonLabel: string
  indentSizeLabel: string
  copyJsonLabel: string
  copiedLabel: string
}>